Switched row-column order in test for size of the rData variable
The aggregation of data from across sessions is performed by growing a matrix. In the prior version of this code, the first dimension of the matrix held the number of trials, and this was tested to determine how to append additional trials from subsequent sessions to the growing matrix. At some point, the row-column order of this matrix was switched. Therefore, I had to change the size test to the second dimension to restore the ability to aggregate data from across sessions. It is possible that this will break the analysis of old data sets which were to be aggregated across sessions.

Signed-off-by: melanopsin <melanopsin@onelightpupil.psych.upenn.edu>